Add categorylink component and update background colors
